Add support for "Bind as User" feature
In some cases, the Base DN to authenticate the user provided by the
connection needs to be obtained by an LDAP search. This change makes it
possible to do so in EMQx.

The substitution of variables in the configuration paramters are
performed unconditionally for both the user and acl processing. This was
necessary to remove the hardcoded Base DN generation and still preserve
backward compatibility. All users benefit from this small improvement,
regardless whether they need the "Bind As User" feature.

The substitution processing is not optimised, but the lists are expected
to be very small or empty, thus no major performance degradation is
expected.

In the LDAP cli implementation, an extra bind is introduced, if
`bind_as_user` is set to true. This was necessary to reset the last bind
to the root user before executing the query.

A major consideration was taken to preserve backward compatibility for
users not opting for the feature. No configuration changes are
necessary, but the `priv/eqmx_auth_ldap.schema` needs to be updated to
pick up the proper default values.